LayoutTests/fast/js/dfg-captured-var-get-local-expected.txt

Tests that GetLocals on captured variables aren't eliminated too aggressively.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS foo() is 42
PASS successfullyParsed is true

TEST COMPLETE
